Make the playlist container more compact
Previously, only 19 songs could fit on the screen inside of the player.
Now, 24 songs can fit on the screen. Eventually, when the window can be
resized, it will be able to fit as many as there are pixels on the
user's side. But for now, this is acceptable. There should be a font
size toggle 'built into the OS'.

But it's readable for now. I'm happy with it.