Add audio tests and then take them away. Add other tests
In this update I also added some sample audio files
for use down the line to test the conversion and the
tags function.
I also refactored some code in the audio context
provider to allow for autoplay of audio. It is working
in the desktop browsers: Chrome and Firefox, but have
not tested with Safari nor mobile... We'll see.
Also got testing finally working with Jest. What a
pain in the ass to do with ES6 modules. Well at least
it's ready for action!
Also, I added a prettierrc file for standardized code
format throughout the repository.