Fix media metadata not updating properly and move providers
The navigator's mediasession's metadata value cannot be stringified
into JSON. So it was not working comparing the stringified values of
two metadata.
Now, if checks if the object has all of the keys of a mediametadata
(because Node.js doesn't have a definition for a MediaMetadata as
there is no GUI) and then compares their values.